# docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/points": {
      "post": {
        "operationId": "PointsController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "point": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer",
                          "example": 1
                        },
                        "studentId": {
                          "type": "integer",
                          "example": 1
                        },
                        "teacherId": {
                          "type": "integer",
                          "example": 1
                        },
                        "reasonId": {
                          "type": "integer",
                          "example": 1
                        },
                        "point": {
                          "type": "integer",
                          "example": 1
                        },
                        "comment": {
                          "type": "string",
                          "example": "string"
                        },
                        "baseDate": {
                          "type": "string",
                          "format": "date-time",
                          "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$",
                          "example": "2026-01-01T00:00:00.000Z"
                        },
                        "updatedDate": {
                          "type": "string",
                          "format": "date-time",
                          "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$",
                          "example": "2026-01-01T00:00:00.000Z"
                        },
                        "student": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "integer",
                              "example": 1
                            },
                            "stuid": {
                              "type": "integer",
                              "example": 1
                            },
                            "name": {
                              "type": "string",
                              "example": "string"
                            },
                            "grade": {
                              "type": "integer",
                              "example": 1
                            },
                            "class": {
                              "type": "integer",
                              "example": 1
                            },
                            "num": {
                              "type": "integer",
                              "example": 1
                            },
                            "point": {
                              "type": "integer",
                              "example": 1
                            }
                          },
                          "required": [
                            "id",
                            "stuid",
                            "name",
                            "grade",
                            "class",
                            "num",
                            "point"
                          ],
                          "additionalProperties": false,
                          "example": {
                            "id": 1,
                            "stuid": 1,
                            "name": "string",
                            "grade": 1,
                            "class": 1,
                            "num": 1,
                            "point": 1
                          }
                        },
                        "teacher": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "integer",
                              "example": 1
                            },
                            "stuid": {
                              "type": "integer",
                              "example": 1
                            },
                            "name": {
                              "type": "string",
                              "example": "string"
                            },
                            "phoneNumber": {
                              "anyOf": [
                                {
                                  "type": "string",
                                  "example": "string"
                                },
                                {
                                  "type": "null",
                                  "example": null
                                }
                              ],
                              "example": "string"
                            },
                            "studentId": {
                              "anyOf": [
                                {
                                  "type": "integer",
                                  "example": 1
                                },
                                {
                                  "type": "null",
                                  "example": null
                                }
                              ],
                              "example": 1
                            }
                          },
                          "required": [
                            "id",
                            "stuid",
                            "name",
                            "phoneNumber",
                            "studentId"
                          ],
                          "additionalProperties": false,
                          "example": {
                            "id": 1,
                            "stuid": 1,
                            "name": "string",
                            "phoneNumber": "string",
                            "studentId": 1
                          }
                        }
                      },
                      "required": [
                        "id",
                        "studentId",
                        "teacherId",
                        "reasonId",
                        "point",
                        "comment",
                        "baseDate",
                        "updatedDate",
                        "student",
                        "teacher"
                      ],
                      "additionalProperties": false,
                      "example": {
                        "id": 1,
                        "studentId": 1,
                        "teacherId": 1,
                        "reasonId": 1,
                        "point": 1,
                        "comment": "string",
                        "baseDate": "2026-01-01T00:00:00.000Z",
                        "updatedDate": "2026-01-01T00:00:00.000Z",
                        "student": {
                          "id": 1,
                          "stuid": 1,
                          "name": "string",
                          "grade": 1,
                          "class": 1,
                          "num": 1,
                          "point": 1
                        },
                        "teacher": {
                          "id": 1,
                          "stuid": 1,
                          "name": "string",
                          "phoneNumber": "string",
                          "studentId": 1
                        }
                      }
                    }
                  },
                  "required": [
                    "point"
                  ],
                  "additionalProperties": false,
                  "example": {
                    "point": {
                      "id": 1,
                      "studentId": 1,
                      "teacherId": 1,
                      "reasonId": 1,
                      "point": 1,
                      "comment": "string",
                      "baseDate": "2026-01-01T00:00:00.000Z",
                      "updatedDate": "2026-01-01T00:00:00.000Z",
                      "student": {
                        "id": 1,
                        "stuid": 1,
                        "name": "string",
                        "grade": 1,
                        "class": 1,
                        "num": 1,
                        "point": 1
                      },
                      "teacher": {
                        "id": 1,
                        "stuid": 1,
                        "name": "string",
                        "phoneNumber": "string",
                        "studentId": 1
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Points"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "studentId": {
                    "type": "integer",
                    "minimum": 1,
                    "example": 1
                  },
                  "reasonId": {
                    "type": "integer",
                    "minimum": 1,
                    "example": 1
                  },
                  "point": {
                    "type": "integer",
                    "minimum": 0,
                    "example": 0
                  },
                  "comment": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 255,
                    "example": "a"
                  }
                },
                "required": [
                  "studentId",
                  "reasonId",
                  "point",
                  "comment"
                ],
                "additionalProperties": false,
                "example": {
                  "studentId": 1,
                  "reasonId": 1,
                  "point": 0,
                  "comment": "a"
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "PointsController_findAll",
        "parameters": [
          {
            "name": "stuId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "example": 1
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "points": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer",
                            "example": 1
                          },
                          "studentId": {
                            "type": "integer",
                            "example": 1
                          },
                          "teacherId": {
                            "type": "integer",
                            "example": 1
                          },
                          "reasonId": {
                            "type": "integer",
                            "example": 1
                          },
                          "point": {
                            "type": "integer",
                            "example": 1
                          },
                          "comment": {
                            "type": "string",
                            "example": "string"
                          },
                          "baseDate": {
                            "type": "string",
                            "format": "date-time",
                            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$",
                            "example": "2026-01-01T00:00:00.000Z"
                          },
                          "updatedDate": {
                            "type": "string",
                            "format": "date-time",
                            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$",
                            "example": "2026-01-01T00:00:00.000Z"
                          },
                          "student": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "integer",
                                "example": 1
                              },
                              "stuid": {
                                "type": "integer",
                                "example": 1
                              },
                              "name": {
                                "type": "string",
                                "example": "string"
                              },
                              "grade": {
                                "type": "integer",
                                "example": 1
                              },
                              "class": {
                                "type": "integer",
                                "example": 1
                              },
                              "num": {
                                "type": "integer",
                                "example": 1
                              },
                              "point": {
                                "type": "integer",
                                "example": 1
                              }
                            },
                            "required": [
                              "id",
                              "stuid",
                              "name",
                              "grade",
                              "class",
                              "num",
                              "point"
                            ],
                            "additionalProperties": false,
                            "example": {
                              "id": 1,
                              "stuid": 1,
                              "name": "string",
                              "grade": 1,
                              "class": 1,
                              "num": 1,
                              "point": 1
                            }
                          },
                          "teacher": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "integer",
                                "example": 1
                              },
                              "stuid": {
                                "type": "integer",
                                "example": 1
                              },
                              "name": {
                                "type": "string",
                                "example": "string"
                              },
                              "phoneNumber": {
                                "anyOf": [
                                  {
                                    "type": "string",
                                    "example": "string"
                                  },
                                  {
                                    "type": "null",
                                    "example": null
                                  }
                                ],
                                "example": "string"
                              },
                              "studentId": {
                                "anyOf": [
                                  {
                                    "type": "integer",
                                    "example": 1
                                  },
                                  {
                                    "type": "null",
                                    "example": null
                                  }
                                ],
                                "example": 1
                              }
                            },
                            "required": [
                              "id",
                              "stuid",
                              "name",
                              "phoneNumber",
                              "studentId"
                            ],
                            "additionalProperties": false,
                            "example": {
                              "id": 1,
                              "stuid": 1,
                              "name": "string",
                              "phoneNumber": "string",
                              "studentId": 1
                            }
                          }
                        },
                        "required": [
                          "id",
                          "studentId",
                          "teacherId",
                          "reasonId",
                          "point",
                          "comment",
                          "baseDate",
                          "updatedDate",
                          "student",
                          "teacher"
                        ],
                        "additionalProperties": false,
                        "example": {
                          "id": 1,
                          "studentId": 1,
                          "teacherId": 1,
                          "reasonId": 1,
                          "point": 1,
                          "comment": "string",
                          "baseDate": "2026-01-01T00:00:00.000Z",
                          "updatedDate": "2026-01-01T00:00:00.000Z",
                          "student": {
                            "id": 1,
                            "stuid": 1,
                            "name": "string",
                            "grade": 1,
                            "class": 1,
                            "num": 1,
                            "point": 1
                          },
                          "teacher": {
                            "id": 1,
                            "stuid": 1,
                            "name": "string",
                            "phoneNumber": "string",
                            "studentId": 1
                          }
                        }
                      },
                      "example": [
                        {
                          "id": 1,
                          "studentId": 1,
                          "teacherId": 1,
                          "reasonId": 1,
                          "point": 1,
                          "comment": "string",
                          "baseDate": "2026-01-01T00:00:00.000Z",
                          "updatedDate": "2026-01-01T00:00:00.000Z",
                          "student": {
                            "id": 1,
                            "stuid": 1,
                            "name": "string",
                            "grade": 1,
                            "class": 1,
                            "num": 1,
                            "point": 1
                          },
                          "teacher": {
                            "id": 1,
                            "stuid": 1,
                            "name": "string",
                            "phoneNumber": "string",
                            "studentId": 1
                          }
                        }
                      ]
                    }
                  },
                  "required": [
                    "points"
                  ],
                  "additionalProperties": false,
                  "example": {
                    "points": [
                      {
                        "id": 1,
                        "studentId": 1,
                        "teacherId": 1,
                        "reasonId": 1,
                        "point": 1,
                        "comment": "string",
                        "baseDate": "2026-01-01T00:00:00.000Z",
                        "updatedDate": "2026-01-01T00:00:00.000Z",
                        "student": {
                          "id": 1,
                          "stuid": 1,
                          "name": "string",
                          "grade": 1,
                          "class": 1,
                          "num": 1,
                          "point": 1
                        },
                        "teacher": {
                          "id": 1,
                          "stuid": 1,
                          "name": "string",
                          "phoneNumber": "string",
                          "studentId": 1
                        }
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Points"
        ]
      }
    },
...
    "/api/points/students": {
      "get": {
        "operationId": "PointsController_findStudents",
        "parameters": [
          {
            "name": "grade",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "example": 1
            }
          },
          {
            "name": "class",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "example": 1
            }
          },
          {
            "name": "number",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "example": 1
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "students": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer",
                            "example": 1
                          },
                          "stuid": {
                            "type": "integer",
                            "example": 1
                          },
                          "name": {
                            "type": "string",
                            "example": "string"
                          },
                          "grade": {
                            "type": "integer",
                            "example": 1
                          },
                          "class": {
                            "type": "integer",
                            "example": 1
                          },
                          "num": {
                            "type": "integer",
                            "example": 1
                          },
                          "point": {
                            "type": "integer",
                            "example": 1
                          }
                        },
                        "required": [
                          "id",
                          "stuid",
                          "name",
                          "grade",
                          "class",
                          "num",
                          "point"
                        ],
                        "additionalProperties": false,
                        "example": {
                          "id": 1,
                          "stuid": 1,
                          "name": "string",
                          "grade": 1,
                          "class": 1,
                          "num": 1,
                          "point": 1
                        }
                      },
                      "example": [
                        {
                          "id": 1,
                          "stuid": 1,
                          "name": "string",
                          "grade": 1,
                          "class": 1,
                          "num": 1,
                          "point": 1
                        }
                      ]
                    }
                  },
                  "required": [
                    "students"
                  ],
                  "additionalProperties": false,
                  "example": {
                    "students": [
                      {
                        "id": 1,
                        "stuid": 1,
                        "name": "string",
                        "grade": 1,
                        "class": 1,
                        "num": 1,
                        "point": 1
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Points"
        ]
      }
    },
...
  },
  "info": {
    "title": "PLMA API",
    "description": "PLMA API documentation",
    "version": "0.0",
    "contact": {}
  },
...
  "servers": [],
...
}